+++ ciphering_cli/src/errors/file_error.js
