+++ ciphering_cli/src/errors/incorrect_config_error.js
