+++ ciphering_cli/src/errors/index.js
