+++ ciphering_cli/src/readable/ciphering_read_stream.js
