+++ ciphering_cli/src/transforms/ciphering_transform.js
