+++ ciphering_cli/src/writable/ciphering_write_stream.js
